+++ internal/cmd/root.go
@@ -24,6 +24,7 @@ func NewRootCommand() *cobra.Command {
 	rootCmd.AddCommand(newVersionCommand())
 	rootCmd.AddCommand(NewCheckCommand())
 	rootCmd.AddCommand(NewVerifyCommand())
+	rootCmd.AddCommand(NewConfigureCommand())
 	rootCmd.AddCommand(newContextCommand())
 
 	return rootCmd